+++ .gitignore
@@ -1,0 +1,5 @@
+node_modules/
+db/
+bundle/
+bundle.zip
+.env
