+++ package.json
@@ -1,0 +1,38 @@
+{
+  "name": "jovo-sample-voice-app-nodejs",
+  "version": "2.0.2",
+  "description": "A sample voice app that works with the Jovo Framework",
+  "main": "src/index.js",
+  "dependencies": {
+    "jovo-cli": "^2.2.3",
+    "jovo-db-filedb": "^2.0.2",
+    "jovo-framework": "^2.2.2",
+    "jovo-platform-alexa": "^2.0.2",
+    "jovo-platform-googleassistant": "^2.0.2",
+    "jovo-plugin-debugger": "^2.0.2"
+  },
+  "devDependencies": {
+    "gulp": "^4.0.0",
+    "gulp-install": "^1.1.0",
+    "gulp-run-command": "0.0.9",
+    "gulp-zip": "^4.2.0",
+    "jest": "^23.6.0"
+  },
+  "scripts": {
+    "tsc": "node -v",
+    "test": "jest",
+    "bundle": "gulp --gulpfile node_modules/jovo-framework/gulpfile.js --cwd ./",
+    "start": "cd src && node index.js --webhook",
+    "launch": "npm start -- --launch"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/jovotech/jovo-sample-voice-app-nodejs.git"
+  },
+  "author": "Jovo",
+  "license": "Apache-2.0",
+  "bugs": {
+    "url": "https://github.com/jovotech/jovo-sample-voice-app-nodejs/issues"
+  },
+  "homepage": "https://github.com/jovotech/jovo-sample-voice-app-nodejs#readme"
+}
