+++ project.js
@@ -1,0 +1,14 @@
+// ------------------------------------------------------------------
+// JOVO PROJECT CONFIGURATION
+// ------------------------------------------------------------------
+
+module.exports = {
+    alexaSkill: {
+       nlu: 'alexa',
+    },
+    googleAction: {
+       nlu: 'dialogflow',
+    },
+    endpoint: '${JOVO_WEBHOOK_URL}',
+};
+
