--- speech-assets/intent-schema.json
@@ -1,22 +1,0 @@
-{
-  "intents": [
-    {
-      "intent": "AMAZON.CancelIntent"
-    },
-    {
-      "intent": "AMAZON.HelpIntent"
-    },
-    {
-      "intent": "AMAZON.StopIntent"
-    },
-    {
-      "slots": [
-        {
-          "name": "CHORD",
-          "type": "ukuleleChords"
-        }
-      ],
-      "intent": "getFingers"
-    }
-  ]
-}
