--- speech-assets/sample-utterances.txt
@@ -1,14 +1,0 @@
-getFingers what is the position for {CHORD} chord
-getFingers what is the fingering for {CHORD} chord
-getFingers how do I play the {CHORD} chord
-getFingers how do I play a {CHORD} chord
-getFingers how do you play a {CHORD} chord
-getFingers how do you play the {CHORD} chord
-getFingers how do you play an {CHORD} chord
-getFingers how do I play an {CHORD} chord
-getFingers I'd like to learn a chord
-getFingers How do I play this chord
-getFingers What's the position for this chord
-getFingers What is the position for this chord
-getFingers Where do I put my fingers for this chord
-getFingers Where do I put my fingers
