+++ src/config.js
@@ -1,0 +1,18 @@
+// ------------------------------------------------------------------
+// APP CONFIGURATION
+// ------------------------------------------------------------------
+
+module.exports = {
+    logging: true,
+ 
+    intentMap: {
+       'AMAZON.StopIntent': 'END',
+    },
+ 
+    db: {
+         FileDb: {
+             pathToFile: '../db/db.json',
+         }
+     },
+ };
+
