+++ test/sample.test.js
@@ -1,0 +1,22 @@
+'use strict';
+const { App, Util } = require('jovo-framework');
+const { GoogleAssistant } = require('jovo-platform-googleassistant');
+const { Alexa } = require('jovo-platform-alexa');
+jest.setTimeout(500);
+
+for (const p of [new Alexa(), new GoogleAssistant()]) {
+    const testSuite = p.makeTestSuite();
+
+    describe(`PLATFORM: ${p.constructor.name} INTENTS` , () => {
+        test('should return a welcome message and ask for the name at "LAUNCH"', async () => {
+            const conversation = testSuite.conversation();
+
+            const launchRequest = await testSuite.requestBuilder.launch();
+            const responseLaunchRequest = await conversation.send(launchRequest);
+            expect(
+                responseLaunchRequest.isAsk('Hello World! What\'s your name?', 'Please tell me your name.')
+            ).toBe(true);
+
+        });
+    });
+}
